requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
openai>=1.3.0
Pillow>=10.0.0
imagehash>=4.3.0
python-dotenv>=1.0.0
supabase>=2.3.0
